requirements.txt
pyaudio
pocketsphinx
python-dotenv
py3-tts
SpeechRecognition
wikipedia
wyze_sdk
# Silences needless ALSA errors  https://github.com/Uberi/speech_recognition/issues/182#issuecomment-1426939447
sounddevice